[수정] 클라이언트 컴포넌트 분리, Suspense 감싸기 구조 변경
